+++ Assets/OkapiKit/Scripts/Variables/Variable.cs
@@ -112,6 +112,5 @@ namespace OkapiKit
 
             return _explanation;
         }
-
     }
 }